pyproject.toml
[tool.poetry]
name = "python-tg-bot"
version = "1.0.0"
description = "Telegram bot for searching books"
authors = ["Shahll <130846011+Shahll@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.0.1"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
logging = "^0.4.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
